+++ packages/atomic/src/components/atomic-search-interface/analytics-config.ts
@@ -4,40 +4,99 @@ import {
 } from '@coveo/headless';
 import {AnalyticsClientSendEventHook} from '@coveo/headless/node_modules/coveo.analytics';
 import {getAtomicEnvironment} from '../../global/environment';
+import {ObservableMap} from '@stencil/store';
+import {AtomicStore, getAllFacets} from '../../utils/store';
 
 type AnalyticsPayload = Parameters<AnalyticsClientSendEventHook>[1];
 
 export function getAnalyticsConfig(
   searchEngineConfig: SearchEngineConfiguration,
-  enabled: boolean
+  enabled: boolean,
+  store: ObservableMap<AtomicStore>
 ): AnalyticsConfiguration {
+  const analyticsClientMiddleware = (
+    event: string,
+    payload: AnalyticsPayload
+  ) => augmentAnalytics(event, payload, store, searchEngineConfig);
+
+  const defaultConfiguration: AnalyticsConfiguration = {
+    analyticsClientMiddleware,
+    enabled,
+    documentLocation: document.location.href,
+    ...(document.referrer && {originLevel3: document.referrer}),
+  };
+
   if (searchEngineConfig.analytics) {
     return {
+      ...defaultConfiguration,
       ...searchEngineConfig.analytics,
-      enabled,
-      analyticsClientMiddleware: (event, payload) =>
-        augmentAnalyticsWithAtomicVersion(
-          event,
-          payload,
-          searchEngineConfig.analytics?.analyticsClientMiddleware
-        ),
     };
   }
-
-  return {
-    enabled,
-    analyticsClientMiddleware: augmentAnalyticsWithAtomicVersion,
-  };
+  return defaultConfiguration;
 }
 
-function augmentAnalyticsWithAtomicVersion(
+function augmentAnalytics(
   event: string,
   payload: AnalyticsPayload,
-  existingMiddleware?: AnalyticsConfiguration['analyticsClientMiddleware']
+  store: ObservableMap<AtomicStore>,
+  config: SearchEngineConfiguration
 ) {
-  const out = existingMiddleware ? existingMiddleware(event, payload) : payload;
-  if (out.customData) {
-    out.customData.coveoAtomicVersion = getAtomicEnvironment().version;
+  let result = augmentWithExternalMiddleware(event, payload, config);
+  result = augmentAnalyticsWithAtomicVersion(result);
+  result = augmentAnalyticsWithFacetTitles(result, store);
+  return result;
+}
+
+function augmentWithExternalMiddleware(
+  event: string,
+  payload: AnalyticsPayload,
+  config: SearchEngineConfiguration
+) {
+  if (config.analytics?.analyticsClientMiddleware) {
+    return config.analytics.analyticsClientMiddleware(event, payload);
   }
-  return out;
+  return payload;
+}
+
+function augmentAnalyticsWithAtomicVersion(payload: AnalyticsPayload) {
+  if (payload.customData) {
+    payload.customData.coveoAtomicVersion = getAtomicEnvironment().version;
+  }
+  return payload;
+}
+
+function augmentAnalyticsWithFacetTitles(
+  payload: AnalyticsPayload,
+  store: ObservableMap<AtomicStore>
+) {
+  const allFacets = getAllFacets(store);
+  const getAtomicFacetLabelOrOriginalTitle = (
+    facetId: string,
+    originalTitle: string
+  ) => (allFacets[facetId] ? allFacets[facetId].label : originalTitle);
+
+  if (payload.facetState) {
+    payload.facetState = payload.facetState.map(
+      (analyticsFacetState: {id: string; title: string}) => {
+        const {id, title: originalTitle} = analyticsFacetState;
+        return {
+          ...analyticsFacetState,
+          title: getAtomicFacetLabelOrOriginalTitle(id, originalTitle),
+        };
+      }
+    );
+  }
+
+  if (
+    payload.customData &&
+    payload.customData.facetTitle &&
+    payload.customData.facetId &&
+    payload.customData.facetTitle
+  ) {
+    payload.customData.facetTitle = getAtomicFacetLabelOrOriginalTitle(
+      payload.customData.facetId,
+      payload.customData.facetTitle
+    );
+  }
+  return payload;
 }